Move classes to its own module
